Turn compression size and scrambled flag into a struct

+++ src/resource/partition_manager.rs
@@ -260,7 +260,7 @@ impl PartitionManager {
 
             let size = |info: &ResourceInfo| {
                 info.compressed_size()
-                    .unwrap_or(info.header.data_size as usize)
+                    .unwrap_or(info.header.data_size)
             };
 
             let changes = partition.resource_patch_indices(rrid);
